requirements.txt
python-slugify
google-genai
playwright
python-dotenv